requirements.txt
langchain
langchain-community
langchain_ollama
langchain-chroma

unstructured
argparse
chromadb
python-dotenv
pypdf
python-multipart

fastapi
uvicorn

# `pip install "unstructured[md]"` 